Simplification and improvments of the pulsing functionality
This change is getting rid of the redundant pulse_* methods. The pulsation
is popssible to control through the 'pulse' parameter on the object and/or
on the method level.

The pulsation loop was modified to allow different styles of pulsation.
Currently it's possible to pulse in the direction UP, DOWN, or BOTH ways.
The direction can be choosen through the 'pulse_dir' parameter on the
object and/or on the method level.

The '__pulse_loop' method was modified to have constant number of steps
per second.

The 'set_leds' method now supports pulsation as well.

The 'led' method now accepts also a list of LEDs.