Fix an issue with time without microseconds
When the time was at microsecond = 0, an error could occur when calling the split function on the second part of the time string.
Instead of spliting the time string, use the attribute of the object datetime (seems easier) and remove unused microseconds.

Another goal was to reduce the cyclomatic complexity of the script (and also have some fun with python, but shh!). Use some of the nice python features and remove the repeated if ... then ... clauses in led method usages.